+++ lib/otp.dart
@@ -2,12 +2,13 @@ library otp;
 
 import 'dart:math';
 import 'dart:typed_data';
-import 'dart:convert';
 import 'package:base32/base32.dart';
 import 'package:convert/convert.dart';
 import 'package:crypto/crypto.dart';
+import 'package:quick_log/quick_log.dart';
 
 class OTP {
+  static bool useTOTPPaddingForHOTP = false;
   /// Generates a Time-based one time password code
   ///
   /// Takes current time in milliseconds, converts to seconds and devides it by interval to get a code every iteration of the interval.
@@ -16,9 +17,12 @@ class OTP {
   /// Optional parameters to change the length of the code provided (default 6), interval (default 30), and hashing algorithm (default SHA1)
   /// These settings are defaulted to the RFC standard but can be changed.
   static int generateTOTPCode(String secret, int time,
-      {int length = 6, int interval = 30, Algorithm algorithm = Algorithm.SHA256}) {
+      {int length = 6,
+      int interval = 30,
+      Algorithm algorithm = Algorithm.SHA256}) {
     time = (((time ~/ 1000).round()) ~/ interval).floor();
-    return _generateCode(secret, time, length, getAlgorithm(algorithm));
+    return _generateCode(secret, time, length, getAlgorithm(algorithm),
+        getAlgorithmByteLength(algorithm));
   }
 
   /// Generates a Time-based one time password code and return as a 0 padded string.
@@ -29,8 +33,11 @@ class OTP {
   /// Optional parameters to change the length of the code provided (default 6), interval (default 30), and hashing algorithm (default SHA1)
   /// These settings are defaulted to the RFC standard but can be changed.
   static String generateTOTPCodeString(String secret, int time,
-      {int length = 6, int interval = 30, Algorithm algorithm = Algorithm.SHA256}) {
-    String code = "${generateTOTPCode(secret, time, length: length, interval: interval, algorithm: algorithm)}";
+      {int length = 6,
+      int interval = 30,
+      Algorithm algorithm = Algorithm.SHA256}) {
+    var code =
+        '${generateTOTPCode(secret, time, length: length, interval: interval, algorithm: algorithm)}';
     return code.padLeft(length, '0');
   }
 
@@ -39,8 +46,11 @@ class OTP {
   /// This function does not increment for you.
   /// Optional parameters to change the length of the code provided (default 6) and hashing algorithm (default SHA1)
   /// These settings are defaulted to the RFC standard but can be changed.
-  static int generateHOTPCode(String secret, int counter, {int length = 6, Algorithm algorithm = Algorithm.SHA256}) {
-    return _generateCode(secret, counter, length, getAlgorithm(algorithm));
+  static int generateHOTPCode(String secret, int counter,
+      {int length = 6, Algorithm algorithm = Algorithm.SHA1}) {
+    return _generateCode(secret, counter, length, getAlgorithm(algorithm),
+        getAlgorithmByteLength(algorithm),
+        isHOTP: true);
   }
 
   /// Generates a one time password code based on a counter you provide and increment, returns as a 0 padded string.
@@ -49,23 +59,33 @@ class OTP {
   /// Optional parameters to change the length of the code provided (default 6) and hashing algorithm (default SHA1)
   /// These settings are defaulted to the RFC standard but can be changed.
   static String generateHOTPCodeString(String secret, int counter,
-      {int length = 6, Algorithm algorithm = Algorithm.SHA256}) {
-    String code = "${generateHOTPCode(secret, counter, length: length, algorithm: algorithm)}";
+      {int length = 6, Algorithm algorithm = Algorithm.SHA1}) {
+    var code =
+        '${generateHOTPCode(secret, counter, length: length, algorithm: algorithm)}';
     return code.padLeft(length, '0');
   }
 
-  static int _generateCode(String secret, int time, int length, Hash mac) {
+  static int _generateCode(
+      String secret, int time, int length, Hash mac, int secretbytes,
+      {bool isHOTP = false}) {
     length = (length > 0) ? length : 6;
 
     var secretList = base32.decode(secret);
+
+    if (!isHOTP || useTOTPPaddingForHOTP) {
+      secretList = _padSecret(secretList, secretbytes);
+    } else {
+      _showHOTPWarning(mac);
+    }
+
     var timebytes = _int2bytes(time);
 
     var hmac = Hmac(mac, secretList);
     var digest = hmac.convert(timebytes).bytes;
 
-    int offset = digest[digest.length - 1] & 0x0f;
+    var offset = digest[digest.length - 1] & 0x0f;
 
-    int binary = ((digest[offset] & 0x7f) << 24) |
+    var binary = ((digest[offset] & 0x7f) << 24) |
         ((digest[offset + 1] & 0xff) << 16) |
         ((digest[offset + 2] & 0xff) << 8) |
         (digest[offset + 3] & 0xff);
@@ -73,44 +93,23 @@ class OTP {
     return binary % pow(10, length);
   }
 
-  static String getDigest(String secret, int counter, int length, Hash mac) {
+  static String getInternalDigest(String secret, int counter, int length, Hash mac) {
     length = (length > 0) ? length : 6;
 
     var secretList = base32.decode(secret);
     var timebytes = _int2bytes(counter);
 
-    var arr = new List<int>.filled(8, 0);
-    //final ctrBytes = arr + [(counter >> 8) & 0xff, counter & 0xff];
-    //final ctrBytes = [counter & 0xff, (counter >> 8) & 0xff] + arr;
-    arr[arr.length-1] = 1;
     var hmac = Hmac(mac, secretList);
     var digest = hmac.convert(timebytes).bytes;
 
     return hex.encode(digest);
   }
-  static int _generateCodeHOTP(String secret, int counter, int length, Hash mac) {
-    length = (length > 0) ? length : 6;
-
-    var secretList = base32.decode(secret);
-    var timebytes = _int2bytes(counter);
-
-    var hmac = Hmac(mac, secretList);
-    var digest = hmac.convert(timebytes).bytes;
-
-    int offset = digest[digest.length - 1] & 0x0f;
-
-    int binary = ((digest[offset] & 0x7f) << 24) |
-        ((digest[offset + 1] & 0xff) << 16) |
-        ((digest[offset + 2] & 0xff) << 8) |
-        (digest[offset + 3] & 0xff);
-
-    return binary % pow(10, length);
-  }
 
   /// Allows you to compare 2 codes in constant time, to mitigate timing attacks for secure codes.
   ///
   /// This function takes 2 codes in string format.
-  static bool constantTimeVerification(final String code, final String othercode) {
+  static bool constantTimeVerification(
+      final String code, final String othercode) {
     if (code.length != othercode.length) {
       return false;
     }
@@ -137,25 +136,32 @@ class OTP {
 
   static Uint8List _int2bytes(int long) {
     // we want to represent the input as a 8-bytes array
-    var byteArray = Uint8List(8);
-    //byteArray..buffer.asUint64List()[0] = long
-    for (var index = byteArray.length - 1; index >= 0; index--) {
-      var byte = long & 0xff;
-      byteArray[index] = byte;
-      long = (long - byte) ~/ 256;
+    return Uint8List(8)..buffer.asByteData().setInt64(0, long);
+  }
+
+  static Uint8List _padSecret(Uint8List secret, int length) {
+    if (secret.length == length) return secret;
+
+    var newList = List<int>();
+    for (var i = 0; i * secret.length < length; i++) {
+      newList.addAll(secret);
     }
-    return byteArray;
+
+    return Uint8List.fromList(newList.sublist(0, length));
+  }
+
+  static void _showHOTPWarning(Hash mac) {
+    if (mac is Sha256 || mac is Sha512) {
+      var logger = Logger("otp");
+      logger.warning('Using non-SHA1 hashing with HOTP is not part of the RFC for HOTP and may cause incompatibilities between different library implementatiions. This library attempts to match behavior with other libraries as best it can.');
+    }
   }
 
   /// Gets the Mac for the provided algorithm.
   static Hash getAlgorithm(Algorithm algorithm) {
     switch (algorithm) {
-      case Algorithm.SHA224:
-        return sha224;
       case Algorithm.SHA256:
         return sha256;
-      case Algorithm.SHA384:
-        return sha384;
       case Algorithm.SHA512:
         return sha512;
       case Algorithm.SHA1:
@@ -164,6 +170,20 @@ class OTP {
         return sha256;
     }
   }
+
+  /// Gets the requried byte length for the provided algorithm.
+  static int getAlgorithmByteLength(Algorithm algorithm) {
+    switch (algorithm) {
+      case Algorithm.SHA256:
+        return 32;
+      case Algorithm.SHA512:
+        return 64;
+      case Algorithm.SHA1:
+        return 20;
+      default:
+        return 32; // For SHA256
+    }
+  }
 }
 
-enum Algorithm { SHA1, SHA224, SHA256, SHA384, SHA512 }
+enum Algorithm { SHA1, SHA256, SHA512 }